apps/server-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
cryptography==44.0.0
stripe==11.3.0
python-dotenv==1.0.1
mangum==0.17.0
pg8000==1.31.2
email-validator==2.1.1
